Enum analysis_driver::ast::PatKind

pub enum PatKind {

    Wild,
    Ident(BindingMode, Ident, Option<P<Pat>>),
    Struct(Option<QSelf>, Path, Vec<PatField, Global>, bool),
    TupleStruct(Option<QSelf>, Path, Vec<P<Pat>, Global>),
    Or(Vec<P<Pat>, Global>),
    Path(Option<QSelf>, Path),
    Tuple(Vec<P<Pat>, Global>),
    Box(P<Pat>),
    Ref(P<Pat>, Mutability),
    Lit(P<Expr>),
    Range(Option<P<Expr>>, Option<P<Expr>>, Spanned<RangeEnd>),
    Slice(Vec<P<Pat>, Global>),
    Rest,
    Paren(P<Pat>),
    MacCall(MacCall),

}

All the different flavors of pattern that Rust recognizes.

Variants

Wild

Represents a wildcard pattern (_).

Ident(BindingMode, Ident, Option<P<Pat>>)

A PatKind::Ident may either be a new bound variable (ref mut binding @ OPT_SUBPATTERN), or a unit struct/variant pattern, or a const pattern (in the last two cases the third field must be None). Disambiguation cannot be done with parser alone, so it happens during name resolution.

Tuple Fields of Ident

0: BindingMode1: Ident2: Option<P<Pat>>

Struct(Option<QSelf>, Path, Vec<PatField, Global>, bool)

A struct or struct variant pattern (e.g., Variant {x, y, ..}). The bool is true in the presence of a ...

Tuple Fields of Struct

0: Option<QSelf>1: Path2: Vec<PatField, Global>3: bool

TupleStruct(Option<QSelf>, Path, Vec<P<Pat>, Global>)

A tuple struct/variant pattern (Variant(x, y, .., z)).

Tuple Fields of TupleStruct

0: Option<QSelf>1: Path2: Vec<P<Pat>, Global>

Or(Vec<P<Pat>, Global>)

An or-pattern A | B | C. Invariant: pats.len() >= 2.

Tuple Fields of Or

0: Vec<P<Pat>, Global>

Path(Option<QSelf>, Path)

A possibly qualified path pattern. Unqualified path patterns A::B::C can legally refer to variants, structs, constants or associated constants. Qualified path patterns <A>::B::C/<A as Trait>::B::C can only legally refer to associated constants.

Tuple Fields of Path

0: Option<QSelf>1: Path

Tuple(Vec<P<Pat>, Global>)

A tuple pattern ((a, b)).

Tuple Fields of Tuple

0: Vec<P<Pat>, Global>

Box(P<Pat>)

A box pattern.

Tuple Fields of Box

0: P<Pat>

Ref(P<Pat>, Mutability)

A reference pattern (e.g., &mut (a, b)).

Tuple Fields of Ref

0: P<Pat>1: Mutability

Lit(P<Expr>)

A literal.

Tuple Fields of Lit

0: P<Expr>

Range(Option<P<Expr>>, Option<P<Expr>>, Spanned<RangeEnd>)

A range pattern (e.g., 1...2, 1..2, 1.., ..2, 1..=2, ..=2).

Tuple Fields of Range

0: Option<P<Expr>>1: Option<P<Expr>>2: Spanned<RangeEnd>

Slice(Vec<P<Pat>, Global>)

A slice pattern [a, b, c].

Tuple Fields of Slice

0: Vec<P<Pat>, Global>

Rest

A rest pattern ...

Syntactically it is valid anywhere.

Semantically however, it only has meaning immediately inside:

a slice pattern: [a, .., b],

a binding pattern immediately inside a slice pattern: [a, r @ ..],

a tuple pattern: (a, .., b),

a tuple struct/variant pattern: $path(a, .., b).

In all of these cases, an additional restriction applies, only one rest pattern may occur in the pattern sequences.

Paren(P<Pat>)

Parentheses in patterns used for grouping (i.e., (PAT)).

Tuple Fields of Paren

0: P<Pat>

MacCall(MacCall)

A macro pattern; pre-expansion.

Tuple Fields of MacCall

0: MacCall