Enum analysis_driver::ast::TyKind

pub enum TyKind {

    Slice(P<Ty>),
    Array(P<Ty>, AnonConst),
    Ptr(MutTy),
    Rptr(Option<Lifetime>, MutTy),
    BareFn(P<BareFnTy>),
    Never,
    Tup(Vec<P<Ty>, Global>),
    Path(Option<QSelf>, Path),
    TraitObject(Vec<GenericBound, Global>, TraitObjectSyntax),
    ImplTrait(NodeId, Vec<GenericBound, Global>),
    Paren(P<Ty>),
    Typeof(AnonConst),
    Infer,
    ImplicitSelf,
    MacCall(MacCall),
    Err,
    CVarArgs,

}

The various kinds of type recognized by the compiler.

Variants

Slice(P<Ty>)

A variable-length slice ([T]).

Tuple Fields of Slice

0: P<Ty>

Array(P<Ty>, AnonConst)

A fixed length array ([T; n]).

Tuple Fields of Array

0: P<Ty>1: AnonConst

Ptr(MutTy)

A raw pointer (*const T or *mut T).

Tuple Fields of Ptr

0: MutTy

Rptr(Option<Lifetime>, MutTy)

A reference (&'a T or &'a mut T).

Tuple Fields of Rptr

0: Option<Lifetime>1: MutTy

BareFn(P<BareFnTy>)

A bare function (e.g., fn(usize) -> bool).

Tuple Fields of BareFn

0: P<BareFnTy>

Never

The never type (!).

Tup(Vec<P<Ty>, Global>)

A tuple ((A, B, C, D,...)).

Tuple Fields of Tup

0: Vec<P<Ty>, Global>

Path(Option<QSelf>, Path)

A path (module::module::...::Type), optionally “qualified”, e.g., <Vec<T> as SomeTrait>::SomeType.

Type parameters are stored in the Path itself.

Tuple Fields of Path

0: Option<QSelf>1: Path

TraitObject(Vec<GenericBound, Global>, TraitObjectSyntax)

A trait object type Bound1 + Bound2 + Bound3 where Bound is a trait or a lifetime.

Tuple Fields of TraitObject

0: Vec<GenericBound, Global>1: TraitObjectSyntax

ImplTrait(NodeId, Vec<GenericBound, Global>)

An impl Bound1 + Bound2 + Bound3 type where Bound is a trait or a lifetime.

The NodeId exists to prevent lowering from having to generate NodeIds on the fly, which would complicate the generation of opaque type Foo = impl Trait items significantly.

Tuple Fields of ImplTrait

0: NodeId1: Vec<GenericBound, Global>

Paren(P<Ty>)

No-op; kept solely so that we can pretty-print faithfully.

Tuple Fields of Paren

0: P<Ty>

Typeof(AnonConst)

Unused for now.

Tuple Fields of Typeof

0: AnonConst

Infer

This means the type should be inferred instead of it having been specified. This can appear anywhere in a type.

ImplicitSelf

Inferred type of a self or &self argument in a method.

MacCall(MacCall)

A macro in the type position.

Tuple Fields of MacCall

0: MacCall

Err

Placeholder for a kind that has failed to be defined.

CVarArgs

Placeholder for a va_list.