Struct analysis_driver::Config

pub(crate) struct Config {

    pub opts: Options,
    pub crate_cfg: HashSet<(String, Option<String>), BuildHasherDefault<FxHasher>>,
    pub input: Input,
    pub input_path: Option<PathBuf>,
    pub output_dir: Option<PathBuf>,
    pub output_file: Option<PathBuf>,
    pub file_loader: Option<Box<dyn FileLoader + Send + Sync + 'static, Global>>,
    pub diagnostic_output: DiagnosticOutput,
    pub stderr: Option<Arc<Mutex<Vec<u8, Global>>>>,
    pub lint_caps: HashMap<LintId, Level, BuildHasherDefault<FxHasher>>,
    pub parse_sess_created: Option<Box<dyn FnOnce(&mut ParseSess) + Send + 'static, Global>>,
    pub register_lints: Option<Box<dyn Fn(&Session, &mut LintStore) + Send + Sync + 'static, Global>>,
    pub override_queries: Option<fn(&Session, &mut Providers, &mut ExternProviders)>,
    pub make_codegen_backend: Option<Box<dyn FnOnce(&Options) + Send + 'static, Global>>,
    pub registry: Registry,

}

The compiler configuration

Fields
opts: Options

Command line options

crate_cfg: HashSet<(String, Option<String>), BuildHasherDefault<FxHasher>>

cfg! configuration in addition to the default ones

input: Inputinput_path: Option<PathBuf>output_dir: Option<PathBuf>output_file: Option<PathBuf>file_loader: Option<Box<dyn FileLoader + Send + Sync + 'static, Global>>diagnostic_output: DiagnosticOutputstderr: Option<Arc<Mutex<Vec<u8, Global>>>>

Set to capture stderr output during compiler execution

lint_caps: HashMap<LintId, Level, BuildHasherDefault<FxHasher>>parse_sess_created: Option<Box<dyn FnOnce(&mut ParseSess) + Send + 'static, Global>>

This is a callback from the driver that is called when [ParseSess] is created.

register_lints: Option<Box<dyn Fn(&Session, &mut LintStore) + Send + Sync + 'static, Global>>

This is a callback from the driver that is called when we’re registering lints; it is called during plugin registration when we have the LintStore in a non-shared state.

Note that if you find a Some here you probably want to call that function in the new function being registered.

override_queries: Option<fn(&Session, &mut Providers, &mut ExternProviders)>

This is a callback from the driver that is called just after we have populated the list of queries.

The second parameter is local providers and the third parameter is external providers.

make_codegen_backend: Option<Box<dyn FnOnce(&Options) + Send + 'static, Global>>

This is a callback from the driver that is called to create a codegen backend.

registry: Registry

Registry of diagnostics codes.