Struct analysis_driver::Session

pub(crate) struct Session {

    pub target: Target,
    pub host: Target,
    pub opts: Options,
    pub host_tlib_path: Rc<SearchPath>,
    pub target_tlib_path: Rc<SearchPath>,
    pub parse_sess: ParseSess,
    pub sysroot: PathBuf,
    pub local_crate_source_file: Option<PathBuf>,
    pub one_time_diagnostics: Lock<HashSet<(DiagnosticMessageId, Option<Span>, String), BuildHasherDefault<FxHasher>>>,
    crate_types: OnceCell<Vec<CrateType, Global>>,
    pub stable_crate_id: OnceCell<StableCrateId>,
    features: OnceCell<Features>,
    incr_comp_session: OneThread<RefCell<IncrCompSession>>,
    pub cgu_reuse_tracker: CguReuseTracker,
    pub prof: SelfProfilerRef,
    pub perf_stats: PerfStats,
    pub code_stats: CodeStats,
    optimization_fuel: Lock<OptimizationFuel>,
    pub print_fuel: Atomic<u64>,
    pub jobserver: Client,
    pub driver_lint_caps: HashMap<LintId, Level, BuildHasherDefault<FxHasher>>,
    pub ctfe_backtrace: Lock<CtfeBacktrace>,
    miri_unleashed_features: Lock<Vec<(Span, Option<Symbol>), Global>>,
    pub asm_arch: Option<InlineAsmArch>,
    pub target_features: HashSet<Symbol, BuildHasherDefault<FxHasher>>,

}

Represents the data associated with a compilation session for a single crate.

Fields
target: Targethost: Targetopts: Optionshost_tlib_path: Rc<SearchPath>target_tlib_path: Rc<SearchPath>parse_sess: ParseSesssysroot: PathBuflocal_crate_source_file: Option<PathBuf>

The name of the root source file of the crate, in the local file system. None means that there is no source file.

one_time_diagnostics: Lock<HashSet<(DiagnosticMessageId, Option<Span>, String), BuildHasherDefault<FxHasher>>>

Set of (DiagnosticId, Option<Span>, message) tuples tracking (sub)diagnostics that have been set once, but should not be set again, in order to avoid redundantly verbose output (Issue #24690, #44953).

crate_types: OnceCell<Vec<CrateType, Global>>stable_crate_id: OnceCell<StableCrateId>

The stable_crate_id is constructed out of the crate name and all the -C metadata arguments passed to the compiler. Its value forms a unique global identifier for the crate. It is used to allow multiple crates with the same name to coexist. See the rustc_codegen_llvm::back::symbol_names module for more information.

features: OnceCell<Features>incr_comp_session: OneThread<RefCell<IncrCompSession>>cgu_reuse_tracker: CguReuseTracker

Used for incremental compilation tests. Will only be populated if -Zquery-dep-graph is specified.

prof: SelfProfilerRef

Used by -Z self-profile.

perf_stats: PerfStats

Some measurements that are being gathered during compilation.

code_stats: CodeStats

Data about code being compiled, gathered during compilation.

optimization_fuel: Lock<OptimizationFuel>

Tracks fuel info if -zfuel=crate=n is specified.

print_fuel: Atomic<u64>

Always set to zero and incremented so that we can print fuel expended by a crate.

jobserver: Client

Loaded up early on in the initialization of this Session to avoid false positives about a job server in our environment.

driver_lint_caps: HashMap<LintId, Level, BuildHasherDefault<FxHasher>>

Cap lint level specified by a driver specifically.

ctfe_backtrace: Lock<CtfeBacktrace>

Tracks the current behavior of the CTFE engine when an error occurs. Options range from returning the error without a backtrace to returning an error and immediately printing the backtrace to stderr. The Lock is only used by miri to allow setting ctfe_backtrace after analysis when MIRI_BACKTRACE is set. This makes it only apply to miri’s errors and not to all CTFE errors.

miri_unleashed_features: Lock<Vec<(Span, Option<Symbol>), Global>>

This tracks where -Zunleash-the-miri-inside-of-you was used to get around a const check, optionally with the relevant feature gate. We use this to warn about unleashing, but with a single diagnostic instead of dozens that drown everything else in noise.

asm_arch: Option<InlineAsmArch>

Architecture to use for interpreting asm!.

target_features: HashSet<Symbol, BuildHasherDefault<FxHasher>>

Set of enabled features for the current target.