Enum analysis_driver::ty::TyKind

pub enum TyKind<'tcx> {

    Bool,
    Char,
    Int(IntTy),
    Uint(UintTy),
    Float(FloatTy),
    Adt(&'tcx AdtDef, &'tcx List<GenericArg<'tcx>>),
    Foreign(DefId),
    Str,
    Array(&'tcx TyS<'tcx>, &'tcx Const<'tcx>),
    Slice(&'tcx TyS<'tcx>),
    RawPtr(TypeAndMut<'tcx>),
    Ref(&'tcx RegionKind, &'tcx TyS<'tcx>, Mutability),
    FnDef(DefId, &'tcx List<GenericArg<'tcx>>),
    FnPtr(Binder<'tcx, FnSig<'tcx>>),
    Dynamic(&'tcx List<Binder<'tcx, ExistentialPredicate<'tcx>>>, &'tcx RegionKind),
    Closure(DefId, &'tcx List<GenericArg<'tcx>>),
    Generator(DefId, &'tcx List<GenericArg<'tcx>>, Movability),
    GeneratorWitness(Binder<'tcx, &'tcx List<&'tcx TyS<'tcx>>>),
    Never,
    Tuple(&'tcx List<GenericArg<'tcx>>),
    Projection(ProjectionTy<'tcx>),
    Opaque(DefId, &'tcx List<GenericArg<'tcx>>),
    Param(ParamTy),
    Bound(DebruijnIndex, BoundTy),
    Placeholder(Placeholder<BoundVar>),
    Infer(InferTy),
    Error(DelaySpanBugEmitted),

}

Defines the kinds of types.

N.B., if you change this, you’ll probably want to change the corresponding AST structure in rustc_ast/src/ast.rs as well.

Variants

Bool

The primitive boolean type. Written as bool.

Char

The primitive character type; holds a Unicode scalar value (a non-surrogate code point). Written as char.

Int(IntTy)

A primitive signed integer type. For example, i32.

Tuple Fields of Int

0: IntTy

Uint(UintTy)

A primitive unsigned integer type. For example, u32.

Tuple Fields of Uint

0: UintTy

Float(FloatTy)

A primitive floating-point type. For example, f64.

Tuple Fields of Float

0: FloatTy

Adt(&'tcx AdtDef, &'tcx List<GenericArg<'tcx>>)

Algebraic data types (ADT). For example: structures, enumerations and unions.

InternalSubsts here, possibly against intuition, may contain Params. That is, even after substitution it is possible that there are type variables. This happens when the Adt corresponds to an ADT definition and not a concrete use of it.

Tuple Fields of Adt

0: &'tcx AdtDef1: &'tcx List<GenericArg<'tcx>>

Foreign(DefId)

An unsized FFI type that is opaque to Rust. Written as extern type T.

Tuple Fields of Foreign

0: DefId

Str

The pointee of a string slice. Written as str.

Array(&'tcx TyS<'tcx>, &'tcx Const<'tcx>)

An array with the given length. Written as [T; n].

Tuple Fields of Array

0: &'tcx TyS<'tcx>1: &'tcx Const<'tcx>

Slice(&'tcx TyS<'tcx>)

The pointee of an array slice. Written as [T].

Tuple Fields of Slice

0: &'tcx TyS<'tcx>

RawPtr(TypeAndMut<'tcx>)

A raw pointer. Written as *mut T or *const T

Tuple Fields of RawPtr

0: TypeAndMut<'tcx>

Ref(&'tcx RegionKind, &'tcx TyS<'tcx>, Mutability)

A reference; a pointer with an associated lifetime. Written as &'a mut T or &'a T.

Tuple Fields of Ref

0: &'tcx RegionKind1: &'tcx TyS<'tcx>2: Mutability

FnDef(DefId, &'tcx List<GenericArg<'tcx>>)

The anonymous type of a function declaration/definition. Each function has a unique type, which is output (for a function named foo returning an i32) as fn() -> i32 {foo}.

For example the type of bar here:

fn foo() -> i32 { 1 }
let bar = foo; // bar: fn() -> i32 {foo}

Tuple Fields of FnDef

0: DefId1: &'tcx List<GenericArg<'tcx>>

FnPtr(Binder<'tcx, FnSig<'tcx>>)

A pointer to a function. Written as fn() -> i32.

For example the type of bar here:

fn foo() -> i32 { 1 }
let bar: fn() -> i32 = foo;

Tuple Fields of FnPtr

0: Binder<'tcx, FnSig<'tcx>>

Dynamic(&'tcx List<Binder<'tcx, ExistentialPredicate<'tcx>>>, &'tcx RegionKind)

A trait object. Written as dyn for<'b> Trait<'b, Assoc = u32> + Send + 'a.

Tuple Fields of Dynamic

0: &'tcx List<Binder<'tcx, ExistentialPredicate<'tcx>>>1: &'tcx RegionKind

Closure(DefId, &'tcx List<GenericArg<'tcx>>)

The anonymous type of a closure. Used to represent the type of |a| a.

Tuple Fields of Closure

0: DefId1: &'tcx List<GenericArg<'tcx>>

Generator(DefId, &'tcx List<GenericArg<'tcx>>, Movability)

The anonymous type of a generator. Used to represent the type of |a| yield a.

Tuple Fields of Generator

0: DefId1: &'tcx List<GenericArg<'tcx>>2: Movability

GeneratorWitness(Binder<'tcx, &'tcx List<&'tcx TyS<'tcx>>>)

A type representing the types stored inside a generator. This should only appear in GeneratorInteriors.

Tuple Fields of GeneratorWitness

0: Binder<'tcx, &'tcx List<&'tcx TyS<'tcx>>>

Never

The never type !.

Tuple(&'tcx List<GenericArg<'tcx>>)

A tuple type. For example, (i32, bool). Use TyS::tuple_fields to iterate over the field types.

Tuple Fields of Tuple

0: &'tcx List<GenericArg<'tcx>>

Projection(ProjectionTy<'tcx>)

The projection of an associated type. For example, <T as Trait<..>>::N.

Tuple Fields of Projection

0: ProjectionTy<'tcx>

Opaque(DefId, &'tcx List<GenericArg<'tcx>>)

Opaque (impl Trait) type found in a return type. The DefId comes either from

the impl Trait ast::Ty node,

or the type Foo = impl Trait declaration The substitutions are for the generics of the function in question. After typeck, the concrete type can be found in the types map.

Tuple Fields of Opaque

0: DefId1: &'tcx List<GenericArg<'tcx>>

Param(ParamTy)

A type parameter; for example, T in fn f<T>(x: T) {}.

Tuple Fields of Param

0: ParamTy

Bound(DebruijnIndex, BoundTy)

Bound type variable, used only when preparing a trait query.

Tuple Fields of Bound

0: DebruijnIndex1: BoundTy

Placeholder(Placeholder<BoundVar>)

A placeholder type - universally quantified higher-ranked type.

Tuple Fields of Placeholder

0: Placeholder<BoundVar>

Infer(InferTy)

A type variable used during type checking.

Tuple Fields of Infer

0: InferTy

Error(DelaySpanBugEmitted)

A placeholder for a type which could not be computed; this is propagated to avoid useless error messages.

Tuple Fields of Error

0: DelaySpanBugEmitted