Function analysis_driver::ty::fast_reject::simplify_type

pub fn simplify_type(
    tcx: TyCtxt<'_>, 
    ty: &TyS<'_>, 
    can_simplify_params: bool
) -> Option<SimplifiedTypeGen<DefId>>

Tries to simplify a type by dropping type parameters, deref’ing away any reference types, etc. The idea is to get something simple that we can use to quickly decide if two types could unify during method lookup.

If can_simplify_params is false, then we will fail to simplify type parameters entirely. This is useful when those type parameters would be instantiated with fresh type variables, since then we can’t say much about whether two types would unify. Put another way, can_simplify_params should be true if type parameters appear free in ty and false if they are to be considered bound.