Trait analysis_driver::ty::layout::FnAbiOf

pub trait FnAbiOf<'tcx>: FnAbiOfHelpers<'tcx> {
    fn fn_abi_of_fn_ptr(
        &self, 
        sig: Binder<'tcx, FnSig<'tcx>>, 
        extra_args: &'tcx List<&'tcx TyS<'tcx>>
    ) -> Self::FnAbiOfResult { ... }

    fn fn_abi_of_instance(
        &self, 
        instance: Instance<'tcx>, 
        extra_args: &'tcx List<&'tcx TyS<'tcx>>
    ) -> Self::FnAbiOfResult { ... }
}

Blanket extension trait for contexts that can compute FnAbis.

Provided methods

fn fn_abi_of_fn_ptr(
&self,
sig: Binder<'tcx, FnSig<'tcx>>,
extra_args: &'tcx List<&'tcx TyS<'tcx>>
) -> Self::FnAbiOfResult

Compute a FnAbi suitable for indirect calls, i.e. to fn pointers.

NB: this doesn’t handle virtual calls - those should use fn_abi_of_instance instead, where the instance is an InstanceDef::Virtual.

fn fn_abi_of_instance(
&self,
instance: Instance<'tcx>,
extra_args: &'tcx List<&'tcx TyS<'tcx>>
) -> Self::FnAbiOfResult

Compute a FnAbi suitable for declaring/defining an fn instance, and for direct calls to an fn.

NB: that includes virtual calls, which are represented by “direct calls” to an InstanceDef::Virtual instance (of <dyn Trait as Trait>::fn).