Function analysis_driver::ty::print::with_no_trimmed_paths

pub fn with_no_trimmed_paths<F, R>(f: F) -> R where
    F: FnOnce() -> R, 

Prevent path trimming if it is turned on. Path trimming affects Display impl of various rustc types, for example std::vec::Vec would be trimmed to Vec, if no other Vec is found.