Module analysis_driver::ty::relate

Generalized type relating mechanism.

A type relation R relates a pair of values (A, B). A and B are usually types or regions but can be other things. Examples of type relations are subtyping, type equality, etc.

Enums

Cause

Traits

Relate

TypeRelation

Functions

expected_found

expected_found_bool

relate_substs

relate_type_and_mut

super_relate_consts

The main “const relation” routine. Note that this does not handle inference artifacts, so you should filter those out before calling it.

super_relate_tys

The main “type relation” routine. Note that this does not handle inference artifacts, so you should filter those out before calling it.

Type Definitions

RelateResult