Struct analysis_driver::ty::GlobalCtxt

pub struct GlobalCtxt<'tcx> {

    pub arena: &'tcx WorkerLocal<Arena<'tcx>>,
    interners: CtxtInterners<'tcx>,
    pub sess: &'tcx Session,
    pub lint_store: Rc<dyn Any + Sync + Send + 'static>,
    pub dep_graph: DepGraph<DepKind>,
    pub prof: SelfProfilerRef,
    pub types: CommonTypes<'tcx>,
    pub lifetimes: CommonLifetimes<'tcx>,
    pub consts: CommonConsts<'tcx>,
    pub(crate) untracked_resolutions: ResolverOutputs,
    pub(crate) untracked_crate: &'tcx Crate<'tcx>,
    pub on_disk_cache: Option<&'tcx (dyn OnDiskCache<'tcx> + 'tcx)>,
    pub queries: &'tcx (dyn QueryEngine<'tcx> + 'tcx),
    pub query_caches: QueryCaches<'tcx>,
    query_kinds: &'tcx [DepKindStruct],
    pub ty_rcache: Lock<HashMap<CReaderCacheKey, &'tcx TyS<'tcx>, BuildHasherDefault<FxHasher>>>,
    pub pred_rcache: Lock<HashMap<CReaderCacheKey, Predicate<'tcx>, BuildHasherDefault<FxHasher>>>,
    pub selection_cache: Cache<(ConstnessAnd<ParamEnvAnd<'tcx, TraitRef<'tcx>>>, ImplPolarity), Result<Option<SelectionCandidate<'tcx>>, SelectionError<'tcx>>>,
    pub evaluation_cache: Cache<(ParamEnvAnd<'tcx, ConstnessAnd<Binder<'tcx, TraitRef<'tcx>>>>, ImplPolarity), EvaluationResult>,
    crate_name: Symbol,
    pub data_layout: TargetDataLayout,
    stability_interner: Sharded<HashMap<&'tcx Stability, (), BuildHasherDefault<FxHasher>>>,
    const_stability_interner: Sharded<HashMap<&'tcx ConstStability, (), BuildHasherDefault<FxHasher>>>,
    pub(crate) alloc_map: Lock<AllocMap<'tcx>>,
    output_filenames: Arc<OutputFilenames>,

}

Fields
arena: &'tcx WorkerLocal<Arena<'tcx>>interners: CtxtInterners<'tcx>sess: &'tcx Sessionlint_store: Rc<dyn Any + Sync + Send + 'static>

This only ever stores a LintStore but we don’t want a dependency on that type here.

FIXME(Centril): consider dyn LintStoreMarker once we can upcast to Any for some additional type safety.

dep_graph: DepGraph<DepKind>prof: SelfProfilerReftypes: CommonTypes<'tcx>

Common types, pre-interned for your convenience.

lifetimes: CommonLifetimes<'tcx>

Common lifetimes, pre-interned for your convenience.

consts: CommonConsts<'tcx>

Common consts, pre-interned for your convenience.

untracked_resolutions: ResolverOutputs

Output of the resolver.

untracked_crate: &'tcx Crate<'tcx>on_disk_cache: Option<&'tcx (dyn OnDiskCache<'tcx> + 'tcx)>

This provides access to the incremental compilation on-disk cache for query results. Do not access this directly. It is only meant to be used by DepGraph::try_mark_green() and the query infrastructure. This is None if we are not incremental compilation mode

queries: &'tcx (dyn QueryEngine<'tcx> + 'tcx)query_caches: QueryCaches<'tcx>query_kinds: &'tcx [DepKindStruct]ty_rcache: Lock<HashMap<CReaderCacheKey, &'tcx TyS<'tcx>, BuildHasherDefault<FxHasher>>>pred_rcache: Lock<HashMap<CReaderCacheKey, Predicate<'tcx>, BuildHasherDefault<FxHasher>>>selection_cache: Cache<(ConstnessAnd<ParamEnvAnd<'tcx, TraitRef<'tcx>>>, ImplPolarity), Result<Option<SelectionCandidate<'tcx>>, SelectionError<'tcx>>>

Caches the results of trait selection. This cache is used for things that do not have to do with the parameters in scope.

evaluation_cache: Cache<(ParamEnvAnd<'tcx, ConstnessAnd<Binder<'tcx, TraitRef<'tcx>>>>, ImplPolarity), EvaluationResult>

Caches the results of trait evaluation. This cache is used for things that do not have to do with the parameters in scope. Merge this with selection_cache?

crate_name: Symbol

The definite name of the current crate after taking into account attributes, commandline parameters, etc.

data_layout: TargetDataLayout

Data layout specification for the current target.

stability_interner: Sharded<HashMap<&'tcx Stability, (), BuildHasherDefault<FxHasher>>>

#[stable] and #[unstable] attributes

const_stability_interner: Sharded<HashMap<&'tcx ConstStability, (), BuildHasherDefault<FxHasher>>>

#[rustc_const_stable] and #[rustc_const_unstable] attributes

alloc_map: Lock<AllocMap<'tcx>>

Stores memory for globals (statics/consts).

output_filenames: Arc<OutputFilenames>