Struct analysis_driver::ty::Instance

pub struct Instance<'tcx> {
    pub def: InstanceDef<'tcx>,
    pub substs: &'tcx List<GenericArg<'tcx>>,
}

A monomorphized InstanceDef.

Monomorphization happens on-the-fly and no monomorphized MIR is ever created. Instead, this type simply couples a potentially generic InstanceDef with some substs, and codegen and const eval will do all required substitution as they run.

Fields
def: InstanceDef<'tcx>substs: &'tcx List<GenericArg<'tcx>>