Module analysis_driver::ty::util

Miscellaneous type-system utilities that are too small to deserve their own modules.

Structs

AlwaysRequiresDrop

Discr

Enums

ExplicitSelf

Traits

IntTypeExt

Functions

fold_list

needs_drop_components

Returns a list of types such that the given type needs drop if and only if any of the returned types need drop. Returns Err(AlwaysRequiresDrop) if this type always needs drop.

normalize_opaque_types

Normalizes all opaque types in the given value, replacing them with their underlying types.

provide